+++ src/ResourceManager/Network/Commands.Network/ExpressRouteCircuit/Stats/GetAzureExpressRouteCircuitARPTableCommand.cs
@@ -54,9 +54,9 @@ namespace Microsoft.Azure.Commands.Network
              Mandatory = false,		
              HelpMessage = "The PeeringType")]		
          [ValidateSet(		
-            MNM.ExpressRouteCircuitPeeringType.AzurePrivatePeering,		
-            MNM.ExpressRouteCircuitPeeringType.AzurePublicPeering,		
-            MNM.ExpressRouteCircuitPeeringType.MicrosoftPeering,		
+            MNM.ExpressRoutePeeringType.AzurePrivatePeering,		
+            MNM.ExpressRoutePeeringType.AzurePublicPeering,		
+            MNM.ExpressRoutePeeringType.MicrosoftPeering,		
             IgnoreCase = true)]		
          public string PeeringType { get; set; }